+++ src/Regularization/Regularization.jl
@@ -11,3 +11,5 @@ include("SundmanTime.jl")
 include("PerturbedLeviCivita.jl")
 
 include("SegmentHandoff.jl")
+
+include("SegmentComposition.jl")